Apply code suggestions
