+++ spyder/widgets/mixins.py
@@ -37,10 +37,10 @@ from spyder.config.main import CONF
 from spyder.py3compat import is_text_string, to_text_string
 from spyder.utils import encoding, sourcecode, programs
 from spyder.utils.misc import get_error_match
-from spyder.utils.syntaxhighlighters import URI_PATTERNS
 from spyder.widgets.arraybuilder import NumpyArrayDialog
 
 QT55_VERSION = programs.check_version(QT_VERSION, "5.5", ">=")
+
 if QT55_VERSION:
     from qtpy.QtCore import QRegularExpression
 else:
@@ -848,47 +848,6 @@ class BaseEditMixin(object):
 
         return word
 
-    def get_uri_at(self, coordinates):
-        """Return uri and cursor for if uri found at coordinates."""
-        return self.get_pattern_cursor_at(URI_PATTERNS, coordinates)
-
-    def get_pattern_cursor_at(self, pattern, coordinates):
-        """
-        Find pattern located at the line where the coordinate is located.
-
-        This returns the actual match and the cursor that selects the text.
-        """
-        # Check if the pattern is in line
-        line = self.get_line_at(coordinates)
-        match = pattern.search(line)
-        uri = None
-        cursor = None
-
-        while match:
-            start, end = match.span()
-
-            # Get cursor selection if pattern found
-            cursor = self.cursorForPosition(coordinates)
-            cursor.movePosition(QTextCursor.StartOfBlock)
-            line_start_position = cursor.position()
-
-            cursor.setPosition(line_start_position + start, cursor.MoveAnchor)
-            start_rect = self.cursorRect(cursor)
-            cursor.setPosition(line_start_position + end, cursor.MoveAnchor)
-            end_rect = self.cursorRect(cursor)
-            bounding_rect = start_rect.united(end_rect)
-
-            # Check if coordinates are located within the selection rect
-            if bounding_rect.contains(coordinates):
-                uri = line[start:end]
-                cursor.setPosition(line_start_position + start,
-                                   cursor.KeepAnchor)
-                break
-            else:
-                match = pattern.search(line, end)
-
-        return uri, cursor
-
     def get_block_indentation(self, block_nb):
         """Return line indentation (character number)"""
         text = to_text_string(self.document().findBlockByNumber(block_nb).text())
@@ -902,6 +861,7 @@ class BaseEditMixin(object):
         block_start = self.document().findBlock(start)
         block_end = self.document().findBlock(end)
         return sorted([block_start.blockNumber(), block_end.blockNumber()])
+
 
     #------Text selection
     def has_selected_text(self):
